moved filterParams and syncToUrl to filterParams.ts lib file from SchedulerWrapper.tsx (#328)
